agents_service/requirements.txt
# ADK Framework
google-adk>=0.1.0

# Google Cloud AI Platform
google-cloud-aiplatform>=1.40.0
google-generativeai>=0.8.0

# Google Cloud Services
google-cloud-storage>=2.10.0
google-cloud-firestore>=2.13.0

# FastAPI & Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pillow>=10.0.0
requests>=2.31.0
